Adicionando pydantic
