+++ tests/test_api.py
@@ -2,7 +2,7 @@ import pytest
 
 from fastapi.testclient import TestClient
 
-from main import app
+from app.main import app
 
 client = TestClient(app)
 
@@ -29,5 +29,4 @@ def test_pega_um_produto():
         "nome": "Smartphone",
         "descricao": "Um telefone que é inteligente",
         "preco": 1500.0,
-        "disponível": True,
 }
